requirements.txt
ipykernel
jupyter 
langchain_community
langchain_core
openai
python-dotenv
